Add relocalisation count to the tally in results
Given a relocalisation and the nodes for which that relocalisation count
should be split between, this function adds the relocalisation count to
the correct tally in the results dataframe.

When you call the function you can pick which tally to increment by
specifying them in a list with the options 'old', 'certain' and 'new'.
Whether it is a gain or loss is specified by the 'relocalisation_type'
argument.