Ensure that topology can only be found for ancestal nodes
The topology_between_nodes fetches a list of nodes between two given
nodes. The logic in the function only works for looking between a node
and an ancestral node (so not a node in a sister lineage or a child). If
non-ancestral nodes are given, then the function may return erroneous
results. Therefore raise an expception if invalid nodes are given.